+++ ortools/linear_solver/xpress_interface.cc
@@ -26,7 +26,7 @@
 #include "ortools/base/logging.h"
 #include "ortools/base/timer.h"
 #include "ortools/linear_solver/linear_solver.h"
-#include "ortools/linear_solver/xpress_basis_status.h"
+#include "ortools/linear_solver/xpress_interface.h"
 #include "ortools/xpress/environment.h"
 
 #define XPRS_INTEGER 'I'
